D_05.02.24

+++ src/main/java/com/bogoliubova/training_service/entity/Book.java
@@ -9,6 +9,7 @@ import lombok.Setter;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
@@ -35,7 +36,7 @@ public class Book {
     private String author;
 
     @Column(name = "b_price")
-    private double bookPrice;
+    private BigDecimal bookPrice;
 
     @OneToMany(mappedBy = "book", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JsonIgnore
