+++ src/main/java/com/bogoliubova/training_service/service/impl/BookServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -73,7 +74,7 @@ public class BookServiceImpl implements BookService {
             book.setAuthor((String) updates.get("author"));
         }
         if (updates.containsKey("bookPrice")) {
-            book.setBookPrice((Double) updates.get("bookPrice"));
+            book.setBookPrice((BigDecimal) updates.get("bookPrice"));
         }
     }
